clean up commits
