+++ core/issuers.js
@@ -4,45 +4,21 @@
 import harden from '@agoric/harden';
 
 import { insist } from '../util/insist';
-import { makeNatAssay } from './assays';
-import { makeBasicMintController } from './mintController';
 
-function makeMint(
-  description,
-  makeMintController = makeBasicMintController,
-  makeAssay = makeNatAssay,
-  makeUseObj = harden((_issuer, _purseOrPayment) => harden({})),
-  parentIssuer,
-) {
+import { makeBasicConfig } from './config/basicConfig';
+
+function makeMint(description, makeConfig = makeBasicConfig, parentIssuer) {
   insist(description)`\
 Description must be truthy: ${description}`;
 
-  let childIssuer;
-  let childMint;
-
-  function getOrMakeChildMint() {
-    if (childMint === undefined) {
-      childMint = makeMint(
-        description,
-        makeMintController,
-        makeAssay,
-        makeUseObj,
-        issuer,
-      );
-      childIssuer = childMint.getIssuer();
-    }
-    return childMint;
-  }
-
-  function getChildAmount(amount) {
-    getOrMakeChildMint();
-    // quantity is the same, but amounts are different for
-    // different issuers
-    const quantity = assay.quantity(amount);
-    const childAssay = childIssuer.getAssay();
-    const childAmount = childAssay.make(quantity);
-    return childAmount;
-  }
+  const {
+    makeCustomIssuer,
+    makeCustomPayment,
+    makeCustomPurse,
+    makeCustomMint,
+    makeMintController,
+    makeAssay,
+  } = makeConfig(makeMint, description);
 
   // assetSrc is a purse or payment. Return a fresh payment.  One internal
   // function used for both cases, since they are so similar.
@@ -57,7 +33,7 @@ Description must be truthy: ${description}`;
     const oldSrcAmount = srcController.getAmount(assetSrc);
     const newSrcAmount = assay.without(oldSrcAmount, paymentAmount);
 
-    const payment = harden({
+    const corePayment = harden({
       getIssuer() {
         return issuer;
       },
@@ -67,21 +43,11 @@ Description must be truthy: ${description}`;
       getName() {
         return paymentName;
       },
-      getUse() {
-        return makeUseObj(issuer, payment);
-      },
-      getDelegatedUse() {
-        // quantity is the same, but amounts are different for
-        // different issuers
-        const childAmount = getChildAmount(payment.getBalance());
-        const childPurse = childMint.mint(childAmount);
-        const childPayment = childPurse.withdrawAll();
-        return makeUseObj(childIssuer, childPayment);
-      },
-      revokeChildren() {
-        const childAmount = getChildAmount(payment.getBalance());
-        childMint.revoke(childAmount);
-      },
+    });
+
+    const payment = harden({
+      ...corePayment,
+      ...makeCustomPayment(issuer, corePayment),
     });
 
     // ///////////////// commit point //////////////////
@@ -93,7 +59,7 @@ Description must be truthy: ${description}`;
     return payment;
   }
 
-  const issuer = harden({
+  const coreIssuer = harden({
     getLabel() {
       return assay.getLabel();
     },
@@ -132,31 +98,19 @@ Description must be truthy: ${description}`;
     burn(amount, srcPaymentP) {
       // We deposit the alleged payment, rather than just doing a get
       // exclusive on it, in order to consume the usage erights as well.
-      const sinkPurse = issuer.makeEmptyPurse('sink purse');
+      const sinkPurse = coreIssuer.makeEmptyPurse('sink purse');
       return sinkPurse.deposit(amount, srcPaymentP);
     },
 
     burnAll(srcPaymentP) {
-      const sinkPurse = issuer.makeEmptyPurse('sink purse');
+      const sinkPurse = coreIssuer.makeEmptyPurse('sink purse');
       return sinkPurse.depositAll(srcPaymentP);
     },
+  });
 
-    getParentIssuer() {
-      return parentIssuer;
-    },
-    getChildIssuer() {
-      getOrMakeChildMint();
-      return childIssuer;
-    },
-    isDescendantIssuer(allegedDescendant) {
-      if (childIssuer === undefined) {
-        return false;
-      }
-      if (childIssuer === issuer) {
-        return true;
-      }
-      return childIssuer.isDescendantIssuer(allegedDescendant);
-    },
+  const issuer = harden({
+    ...coreIssuer,
+    ...makeCustomIssuer(coreIssuer, parentIssuer),
   });
 
   const label = harden({ issuer, description });
@@ -184,7 +138,7 @@ Description must be truthy: ${description}`;
     return amount;
   }
 
-  const mint = harden({
+  const coreMint = harden({
     getIssuer() {
       return issuer;
     },
@@ -192,28 +146,11 @@ Description must be truthy: ${description}`;
       purseController.destroyAll();
       paymentController.destroyAll();
     },
-    revoke(amount) {
-      amount = assay.coerce(amount);
-      // for non-fungible tokens that are unique, revoke them by removing them from
-      // the purses/payments that they live in
-
-      // the amount may not exist in the case of childMint amounts, so
-      // catch the error
-
-      try {
-        destroy(amount);
-        if (childMint !== undefined) {
-          childMint.revoke(getChildAmount(amount)); // recursively revoke child assets
-        }
-      } catch (err) {
-        console.log(err);
-      }
-    },
     mint(initialBalance, name = 'a purse') {
       initialBalance = assay.coerce(initialBalance);
       name = `${name}`;
 
-      const purse = harden({
+      const corePurse = harden({
         getName() {
           return name;
         },
@@ -248,95 +185,27 @@ Description must be truthy: ${description}`;
             paymentName,
           );
         },
-        getUse() {
-          return makeUseObj(issuer, purse);
-        },
-        getDelegatedUse() {
-          const childAmount = getChildAmount(purse.getBalance());
-          const childPurse = childMint.mint(childAmount);
-          return makeUseObj(childIssuer, childPurse);
-        },
-        revokeChildren() {
-          const childAmount = getChildAmount(purse.getBalance());
-          childMint.revoke(childAmount);
-        },
+      });
+
+      const delegatedUsePurseMethods = makeCustomPurse(issuer, corePurse);
+
+      const purse = harden({
+        ...corePurse,
+        ...delegatedUsePurseMethods,
       });
       purseController.recordNew(purse, initialBalance);
       return purse;
     },
   });
+
+  const customMint = makeCustomMint(assay, destroy);
+
+  const mint = harden({
+    ...coreMint,
+    ...customMint,
+  });
   return mint;
 }
 harden(makeMint);
 
-// Creates a local issuer that locally represents a remotely issued
-// currency. Returns a promise for a peg object that asynchonously
-// converts between the two. The local currency is synchronously
-// transferable locally.
-function makePeg(
-  E,
-  remoteIssuerP,
-  makeMintController,
-  makeAssay = makeNatAssay,
-) {
-  const remoteLabelP = E(remoteIssuerP).getLabel();
-
-  // The remoteLabel is a local copy of the remote pass-by-copy
-  // label. It has a presence of the remote issuer and a copy of the
-  // description.
-  return Promise.resolve(remoteLabelP).then(remoteLabel => {
-    // Retaining remote currency deposits it in here.
-    // Redeeming local currency withdraws remote from here.
-    const backingPurseP = E(remoteIssuerP).makeEmptyPurse('backing');
-
-    const { description } = remoteLabel;
-    const localMint = makeMint(description, makeMintController, makeAssay);
-    const localIssuer = localMint.getIssuer();
-    const localLabel = localIssuer.getLabel();
-
-    function localAmountOf(remoteAmount) {
-      return harden({
-        label: localLabel,
-        quantity: remoteAmount.quantity,
-      });
-    }
-
-    function remoteAmountOf(localAmount) {
-      return harden({
-        label: remoteLabel,
-        quantity: localAmount.quantity,
-      });
-    }
-
-    return harden({
-      getLocalIssuer() {
-        return localIssuer;
-      },
-
-      getRemoteIssuer() {
-        return remoteIssuerP;
-      },
-
-      retainAll(remotePaymentP, name = 'backed') {
-        return E(backingPurseP)
-          .depositAll(remotePaymentP)
-          .then(remoteAmount =>
-            localMint
-              .mint(localAmountOf(remoteAmount), `${name} purse`)
-              .withdrawAll(name),
-          );
-      },
-
-      redeemAll(localPayment, name = 'redeemed') {
-        return localIssuer
-          .burnAll(localPayment)
-          .then(localAmount =>
-            E(backingPurseP).withdraw(remoteAmountOf(localAmount), name),
-          );
-      },
-    });
-  });
-}
-harden(makePeg);
-
-export { makeMint, makePeg };
+export { makeMint };
