+++ more/peg/peg.js
@@ -1,0 +1,75 @@
+import harden from '@agoric/harden';
+import { makeNatAssay } from '../../core/config/assays';
+import { makeMint } from '../../core/issuers';
+
+// Creates a local issuer that locally represents a remotely issued
+// currency. Returns a promise for a peg object that asynchonously
+// converts between the two. The local currency is synchronously
+// transferable locally.
+function makePeg(
+  E,
+  remoteIssuerP,
+  makeMintController,
+  makeAssay = makeNatAssay,
+) {
+  const remoteLabelP = E(remoteIssuerP).getLabel();
+
+  // The remoteLabel is a local copy of the remote pass-by-copy
+  // label. It has a presence of the remote issuer and a copy of the
+  // description.
+  return Promise.resolve(remoteLabelP).then(remoteLabel => {
+    // Retaining remote currency deposits it in here.
+    // Redeeming local currency withdraws remote from here.
+    const backingPurseP = E(remoteIssuerP).makeEmptyPurse('backing');
+
+    const { description } = remoteLabel;
+    const localMint = makeMint(description, makeMintController, makeAssay);
+    const localIssuer = localMint.getIssuer();
+    const localLabel = localIssuer.getLabel();
+
+    function localAmountOf(remoteAmount) {
+      return harden({
+        label: localLabel,
+        quantity: remoteAmount.quantity,
+      });
+    }
+
+    function remoteAmountOf(localAmount) {
+      return harden({
+        label: remoteLabel,
+        quantity: localAmount.quantity,
+      });
+    }
+
+    return harden({
+      getLocalIssuer() {
+        return localIssuer;
+      },
+
+      getRemoteIssuer() {
+        return remoteIssuerP;
+      },
+
+      retainAll(remotePaymentP, name = 'backed') {
+        return E(backingPurseP)
+          .depositAll(remotePaymentP)
+          .then(remoteAmount =>
+            localMint
+              .mint(localAmountOf(remoteAmount), `${name} purse`)
+              .withdrawAll(name),
+          );
+      },
+
+      redeemAll(localPayment, name = 'redeemed') {
+        return localIssuer
+          .burnAll(localPayment)
+          .then(localAmount =>
+            E(backingPurseP).withdraw(remoteAmountOf(localAmount), name),
+          );
+      },
+    });
+  });
+}
+harden(makePeg);
+
+export { makePeg };
