+++ test/swingsetTests/escrow/bootstrap.js
@@ -3,8 +3,7 @@
 import harden from '@agoric/harden';
 
 import { escrowExchangeSrcs } from '../../../core/escrow';
-import { makeUniAssayMaker } from '../../../core/assays';
-import { makeBasicMintController } from '../../../core/mintController';
+import { makeInviteConfigMaker } from '../../../core/config/inviteConfig';
 import { insist } from '../../../util/insist';
 import { makeMint } from '../../../core/issuers';
 import { allComparable } from '../../../util/sameStructure';
@@ -163,11 +162,10 @@ expected successful check ${result}`;
     async bootstrap(argv, vats) {
       const host = await E(vats.host).makeHost();
       const randMintP = E(vats.mint).makeMint('rand');
-      const artMintP = makeMint(
-        'art',
-        makeBasicMintController,
-        makeUniAssayMaker(),
-      );
+
+      const { makeInviteConfig } = makeInviteConfigMaker();
+
+      const artMintP = makeMint('art', makeInviteConfig);
       switch (argv[0]) {
         case 'escrow misMatches': {
           return testEscrowServiceMismatches(host, randMintP, artMintP);
